Add ability to search movies endpoint of TMDB API
When a movie search GET request comes in, the app will
use the search parameter passed in as the query parameter
for an API call to the TMDB /search/movie endpoint.  When the
result of the call comes back, the app returns the json result
to the client.

The TMDB API key is accessed through an environment variable.